requirements.txt
aiohttp==3.8.4
discord.py==2.2.3
sentry-sdk==1.28.1